% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{Covid19EstimationMentaldisorder}
\alias{Covid19EstimationMentaldisorder}
\alias{Covid19EstimationMentaldisorder-package}
\title{Covid19EstimationMentaldisorder}
\description{
Covid19EstimationMentaldisorder
}
